+++ NEWS.md
@@ -3,6 +3,7 @@ Development version
 * switch group and unit definitions, to be consistent with the literature
 * add Theil's Information Index (H) 
 * add entropy function
+* add mutual_within function to decompose weighted within indices
 
 v0.0.1
 ======